+++ src/app/(main)/meeting/participate/feedback/[groupId]/funnels/step2/Step2.tsx
@@ -1,5 +1,5 @@
-import { useFeedbackContext } from './FeedbackProvider.client';
-import TitleSection from './TitleSection';
+import { useFeedbackContext } from '../../components/FeedbackProvider.client';
+import TitleSection from '../../components/TitleSection';
 import { EstimateResponse } from '@/apis/groups';
 import { Avatar } from '@/components/Avatar';
 import { Button, ButtonGroup } from '@/components/Button';
